resolves Ian Comments
