+++ packages/frontend/src/components/Usage/Container.vue
@@ -1,60 +1,40 @@
 <script setup lang="ts">
 import Card from "primevue/card";
-import { onMounted, onUnmounted, ref } from "vue";
 
-import Content from "./Content.vue";
-import Header from "./Header.vue";
-import { useSidebar } from "./useSidebar";
+import { useForm } from "./useForm";
 
-const { sections, activeSection, scrollToSection } = useSidebar();
-const contentRef = ref<HTMLElement>();
-
-const handleScroll = () => {
-  if (contentRef.value === undefined) return;
-  const scrollPosition = contentRef.value.scrollTop + 200;
-  for (let i = sections.length - 1; i >= 0; i--) {
-    const section = sections[i];
-    if (section === undefined) continue;
-    const element = document.getElementById(section.id);
-    if (element !== null && element.offsetTop <= scrollPosition) {
-      activeSection.value = section.id;
-      break;
-    }
-  }
-};
-
-onMounted(() => {
-  contentRef.value?.addEventListener("scroll", handleScroll);
-});
-
-onUnmounted(() => {
-  contentRef.value?.removeEventListener("scroll", handleScroll);
-});
-
-const isActive = (sectionId: string) => activeSection.value === sectionId;
+const {
+  sections,
+  activeSection,
+  contentRef,
+  scrollToSection,
+  handleScroll,
+  isActive,
+} = useForm();
 </script>
 
 <template>
   <div class="h-full flex flex-col gap-1">
     <Card
       class="h-fit"
-      :pt="{
-        body: { class: 'h-fit p-0 flex flex-col' },
-        content: { class: 'h-fit flex flex-col' },
-      }"
+      pt:body:class="h-fit p-0 flex flex-col"
+      pt:content:class="h-fit flex flex-col"
     >
       <template #header>
-        <Header />
+        <div class="p-4">
+          <h2 class="text-lg font-semibold">Usage</h2>
+          <p class="text-sm text-gray-400">
+            Learn how to use Panes to create and view custom transformations
+          </p>
+        </div>
       </template>
     </Card>
 
     <div class="flex gap-1 h-full overflow-auto">
       <Card
         class="h-full w-[220px] flex-shrink-0"
-        :pt="{
-          body: { class: 'h-full p-0 flex flex-col' },
-          content: { class: 'h-full flex flex-col' },
-        }"
+        pt:body:class="h-full p-0 flex flex-col"
+        pt:content:class="h-full flex flex-col"
       >
         <template #content>
           <div class="h-full overflow-auto p-3 space-y-1">
@@ -77,14 +57,26 @@ const isActive = (sectionId: string) => activeSection.value === sectionId;
 
       <Card
         class="h-full flex-1"
-        :pt="{
-          body: { class: 'h-full p-0 flex flex-col' },
-          content: { class: 'h-full flex flex-col' },
-        }"
+        pt:body:class="h-full p-0 flex flex-col"
+        pt:content:class="h-full flex flex-col"
       >
         <template #content>
           <div ref="contentRef" class="h-full overflow-auto p-4">
-            <Content />
+            <div class="space-y-8">
+              <section
+                v-for="section in sections"
+                :id="section.id"
+                :key="section.id"
+                class="space-y-3"
+              >
+                <h2 class="text-2xl font-semibold">{{ section.title }}</h2>
+                <div class="space-y-3 text-surface-300 leading-relaxed">
+                  <p v-for="(para, idx) in section.body" :key="idx">
+                    {{ para }}
+                  </p>
+                </div>
+              </section>
+            </div>
           </div>
         </template>
       </Card>
