--- packages/frontend/src/components/Usage/Header.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-  <div class="p-4">
-    <h2 class="text-lg font-semibold">Usage</h2>
-    <p class="text-sm text-gray-400">
-      Learn how to use Panes to create and view custom transformations
-    </p>
-  </div>
-</template>
-
