+++ packages/frontend/src/components/Usage/index.ts
@@ -1,2 +1,1 @@
 export { default as Usage } from "./Container.vue";
-
